move builders to a separate utils file
